Added flip board hotkey
Oxidation 0.6.2 changes:
More aggressive LMP (13 elo)
Eval tuning (4 elo)
History decay (14 elo STC, 9 elo LTC)
Mobility (about 50 elo in standard-like variants)
King movegen speedups (10 elo VSTC)
16 byte TT entries (7 elo with constrained hash)
Retested Pesto eval, it is currently worth 90 elo

Bench: 9398185

+++ liberty_chess/src/moves.rs
@@ -12,8 +12,8 @@ enum Stage {
 /// A struct to represent a move
 #[derive(Clone, Copy, Eq, PartialEq)]
 pub struct Move {
-  start: (u16, u16),
-  end: (u16, u16),
+  start: (u8, u8),
+  end: (u8, u8),
   promotion: Option<Piece>,
 }
 
@@ -65,8 +65,8 @@ impl FromStr for Move {
                 Err(())
               } else {
                 Ok(Self {
-                  start: (start_row as u16 - 1, start_col as u16 - 1),
-                  end: (end_row as u16 - 1, end_col as u16 - 1),
+                  start: (start_row as u8 - 1, start_col as u8 - 1),
+                  end: (end_row as u8 - 1, end_col as u8 - 1),
                   promotion,
                 })
               };
@@ -94,8 +94,8 @@ impl FromStr for Move {
             Err(())
           } else {
             Ok(Self {
-              start: (start_row as u16 - 1, start_col as u16 - 1),
-              end: (end_row as u16 - 1, end_col as u16 - 1),
+              start: (start_row as u8 - 1, start_col as u8 - 1),
+              end: (end_row as u8 - 1, end_col as u8 - 1),
               promotion: None,
             })
           }
@@ -112,8 +112,8 @@ impl Move {
   #[must_use]
   pub const fn new(start: (usize, usize), end: (usize, usize)) -> Self {
     Self {
-      start: (start.0 as u16, start.1 as u16),
-      end: (end.0 as u16, end.1 as u16),
+      start: (start.0 as u8, start.1 as u8),
+      end: (end.0 as u8, end.1 as u8),
       promotion: None,
     }
   }
